+++ dio/lib/src/dio_mixin.dart
@@ -4,10 +4,12 @@ import 'dart:convert';
 import 'dart:math' as math;
 import 'dart:typed_data';
 
+import 'package:meta/meta.dart';
+
 import 'adapter.dart';
 import 'cancel_token.dart';
 import 'dio.dart';
-import 'dio_error.dart';
+import 'dio_exception.dart';
 import 'form_data.dart';
 import 'headers.dart';
 import 'interceptors/imply_content_type.dart';
@@ -359,8 +361,7 @@ abstract class DioMixin implements Dio {
     ProgressCallback? onSendProgress,
     ProgressCallback? onReceiveProgress,
   }) async {
-    options ??= Options();
-    final requestOptions = options.compose(
+    final requestOptions = (options ?? Options()).compose(
       this.options,
       path,
       data: data,
@@ -368,13 +369,11 @@ abstract class DioMixin implements Dio {
       onReceiveProgress: onReceiveProgress,
       onSendProgress: onSendProgress,
       cancelToken: cancelToken,
+      sourceStackTrace: StackTrace.current,
     );
-    requestOptions.onReceiveProgress = onReceiveProgress;
-    requestOptions.onSendProgress = onSendProgress;
-    requestOptions.cancelToken = cancelToken;
 
     if (_closed) {
-      throw DioError.connectionError(
+      throw DioException.connectionError(
         reason: "Dio can't establish a new connection after it was closed.",
         requestOptions: requestOptions,
       );
@@ -385,8 +384,6 @@ abstract class DioMixin implements Dio {
 
   @override
   Future<Response<T>> fetch<T>(RequestOptions requestOptions) async {
-    requestOptions.cancelToken?.requestOptions = requestOptions;
-
     if (T != dynamic &&
         !(requestOptions.responseType == ResponseType.bytes ||
             requestOptions.responseType == ResponseType.stream)) {
@@ -444,28 +441,26 @@ abstract class DioMixin implements Dio {
 
     // Convert the error interceptor to a functional callback in which
     // we can handle the return value of interceptor callback.
-    FutureOr<dynamic> Function(dynamic, StackTrace) errorInterceptorWrapper(
+    FutureOr<dynamic> Function(Object) errorInterceptorWrapper(
       InterceptorErrorCallback interceptor,
     ) {
-      return (err, stackTrace) {
-        if (err is! InterceptorState) {
-          err = InterceptorState(
-            assureDioError(err, requestOptions, stackTrace),
-          );
-        }
-
+      return (err) {
+        final state = err is InterceptorState
+            ? err
+            : InterceptorState(assureDioException(err, requestOptions));
         Future<InterceptorState> handleError() async {
           final errorHandler = ErrorInterceptorHandler();
-          interceptor(err.data, errorHandler);
+          interceptor(state.data, errorHandler);
           return errorHandler.future;
         }
 
         // The request has already been canceled,
         // there is no need to listen for another cancellation.
-        if (err.data is DioError && err.data.type == DioErrorType.cancel) {
+        if (state.data is DioException &&
+            state.data.type == DioExceptionType.cancel) {
           return handleError();
-        } else if (err.type == InterceptorResultType.next ||
-            err.type == InterceptorResultType.rejectCallFollowing) {
+        } else if (state.type == InterceptorResultType.next ||
+            state.type == InterceptorResultType.rejectCallFollowing) {
           return listenCancelForAsyncTask(
             requestOptions.cancelToken,
             Future(handleError),
@@ -486,8 +481,9 @@ abstract class DioMixin implements Dio {
 
     // Add request interceptors to request flow
     for (final interceptor in interceptors) {
-      final fun =
-          interceptor is QueuedInterceptor ? interceptor._handleRequest : interceptor.onRequest;
+      final fun = interceptor is QueuedInterceptor
+          ? interceptor._handleRequest
+          : interceptor.onRequest;
       future = future.then(requestInterceptorWrapper(fun));
     }
 
@@ -498,38 +494,43 @@ abstract class DioMixin implements Dio {
         RequestInterceptorHandler handler,
       ) {
         requestOptions = reqOpt;
-        _dispatchRequest<T>(reqOpt).then((value) => handler.resolve(value, true)).catchError((e) {
-          handler.reject(e as DioError, true);
+        _dispatchRequest<T>(reqOpt)
+            .then((value) => handler.resolve(value, true))
+            .catchError((e) {
+          handler.reject(e as DioException, true);
         });
       }),
     );
 
     // Add response interceptors to request flow
     for (final interceptor in interceptors) {
-      final fun =
-          interceptor is QueuedInterceptor ? interceptor._handleResponse : interceptor.onResponse;
+      final fun = interceptor is QueuedInterceptor
+          ? interceptor._handleResponse
+          : interceptor.onResponse;
       future = future.then(responseInterceptorWrapper(fun));
     }
 
     // Add error handlers to request flow
     for (final interceptor in interceptors) {
-      final fun = interceptor is QueuedInterceptor ? interceptor._handleError : interceptor.onError;
+      final fun = interceptor is QueuedInterceptor
+          ? interceptor._handleError
+          : interceptor.onError;
       future = future.catchError(errorInterceptorWrapper(fun));
     }
-    // Normalize errors, we convert error to the DioError.
+    // Normalize errors, we convert error to the DioException.
     return future.then<Response<T>>((data) {
       return assureResponse<T>(
         data is InterceptorState ? data.data : data,
         requestOptions,
       );
-    }).catchError((dynamic e, StackTrace s) {
+    }).catchError((Object e) {
       final isState = e is InterceptorState;
       if (isState) {
         if (e.type == InterceptorResultType.resolve) {
           return assureResponse<T>(e.data, requestOptions);
         }
       }
-      throw assureDioError(isState ? e.data : e, requestOptions, s);
+      throw assureDioException(isState ? e.data : e, requestOptions);
     });
   }
 
@@ -577,14 +578,14 @@ abstract class DioMixin implements Dio {
       if (statusOk) {
         return ret;
       } else {
-        throw DioError.badResponse(
+        throw DioException.badResponse(
           statusCode: responseBody.statusCode,
           requestOptions: reqOpt,
           response: ret,
         );
       }
-    } catch (e, stackTrace) {
-      throw assureDioError(e, reqOpt, stackTrace);
+    } catch (e) {
+      throw assureDioException(e, reqOpt);
     }
   }
 
@@ -604,7 +605,8 @@ abstract class DioMixin implements Dio {
         r' ABCDEFGHIJKLMNOPQRSTUVWXYZ   ^_'
         r'`abcdefghijklmnopqrstuvwxyz | ~ ';
     for (final int codeUnit in token.codeUnits) {
-      if (codeUnit >= validChars.length || validChars.codeUnitAt(codeUnit) == 0x20) {
+      if (codeUnit >= validChars.length ||
+          validChars.codeUnitAt(codeUnit) == 0x20) {
         return false;
       }
     }
@@ -637,7 +639,8 @@ abstract class DioMixin implements Dio {
           return false;
         });
       } else if (data is FormData) {
-        options.headers[Headers.contentTypeHeader] = '${Headers.multipartFormDataContentType}; '
+        options.headers[Headers.contentTypeHeader] =
+            '${Headers.multipartFormDataContentType}; '
             'boundary=${data.boundary}';
         stream = data.finalize();
         length = data.length;
@@ -671,6 +674,7 @@ abstract class DioMixin implements Dio {
   }
 
   // If the request has been cancelled, stop request and throw error.
+  @internal
   static void checkCancelled(CancelToken? cancelToken) {
     final error = cancelToken?.cancelError;
     if (error != null) {
@@ -678,6 +682,7 @@ abstract class DioMixin implements Dio {
     }
   }
 
+  @internal
   static Future<T> listenCancelForAsyncTask<T>(
     CancelToken? cancelToken,
     Future<T> future,
@@ -688,36 +693,37 @@ abstract class DioMixin implements Dio {
     ]);
   }
 
+  @internal
   static Options checkOptions(String method, Options? options) {
     options ??= Options();
     options.method = method;
     return options;
   }
 
-  static DioError assureDioError(
+  @internal
+  static DioException assureDioException(
     Object err,
     RequestOptions requestOptions,
-    StackTrace? sourceStackTrace,
   ) {
-    if (err is DioError) {
+    if (err is DioException) {
       // nothing to be done
       return err;
     }
-    return DioError(
+    return DioException(
       requestOptions: requestOptions,
       error: err,
-      stackTrace: sourceStackTrace,
     );
   }
 
+  @internal
   static Response<T> assureResponse<T>(
-    Object response, [
-    RequestOptions? requestOptions,
-  ]) {
+    Object response,
+    RequestOptions requestOptions,
+  ) {
     if (response is! Response) {
       return Response<T>(
         data: response as T,
-        requestOptions: requestOptions ?? RequestOptions(path: ''),
+        requestOptions: requestOptions,
       );
     } else if (response is! Response<T>) {
       final T? data = response.data as T?;