+++ plugins/cookie_manager/lib/src/cookie_mgr.dart
@@ -28,7 +28,20 @@ class CookieManager extends Interceptor {
   final CookieJar cookieJar;
 
   /// Merge cookies into a Cookie string.
+  /// Cookies with longer paths are listed before cookies with shorter paths.
   static String getCookies(List<Cookie> cookies) {
+    // Sort cookies by path (longer path first).
+    cookies.sort((a, b) {
+      if (a.path == null && b.path == null) {
+        return 0;
+      } else if (a.path == null) {
+        return -1;
+      } else if (b.path == null) {
+        return 1;
+      } else {
+        return b.path!.length.compareTo(a.path!.length);
+      }
+    });
     return cookies.map((cookie) => '${cookie.name}=${cookie.value}').join('; ');
   }
 
@@ -38,17 +51,21 @@ class CookieManager extends Interceptor {
       final previousCookies =
           options.headers[HttpHeaders.cookieHeader] as String?;
       final newCookies = getCookies([
-        ...cookies,
         ...?previousCookies
             ?.split(';')
             .where((e) => e.isNotEmpty)
             .map((c) => Cookie.fromSetCookieValue(c)),
+        ...cookies,
       ]);
       options.headers[HttpHeaders.cookieHeader] =
           newCookies.isNotEmpty ? newCookies : null;
       handler.next(options);
     }).catchError((dynamic e, StackTrace s) {
-      final err = DioError(requestOptions: options, error: e, stackTrace: s);
+      final err = DioException(
+        requestOptions: options,
+        error: e,
+        stackTrace: s,
+      );
       handler.reject(err, true);
     });
   }
@@ -57,7 +74,7 @@ class CookieManager extends Interceptor {
   void onResponse(Response response, ResponseInterceptorHandler handler) {
     _saveCookies(response).then((_) => handler.next(response)).catchError(
       (dynamic e, StackTrace s) {
-        final err = DioError(
+        final err = DioException(
           requestOptions: response.requestOptions,
           error: e,
           stackTrace: s,
@@ -68,11 +85,11 @@ class CookieManager extends Interceptor {
   }
 
   @override
-  void onError(DioError err, ErrorInterceptorHandler handler) {
+  void onError(DioException err, ErrorInterceptorHandler handler) {
     if (err.response != null) {
       _saveCookies(err.response!).then((_) => handler.next(err)).catchError(
         (dynamic e, StackTrace s) {
-          final error = DioError(
+          final error = DioException(
             requestOptions: err.response!.requestOptions,
             error: e,
             stackTrace: s,
@@ -105,17 +122,19 @@ class CookieManager extends Interceptor {
     // cookie handling here, because when `followRedirects` is set to false,
     // users will be available to handle cookies themselves.
     final isRedirectRequest = statusCode >= 300 && statusCode < 400;
+    // Saving cookies for the original site.
+    await cookieJar.saveFromResponse(response.realUri, cookies);
     if (isRedirectRequest && locations.isNotEmpty) {
+      final originalUri = response.realUri;
       await Future.wait(
         locations.map(
           (location) => cookieJar.saveFromResponse(
-            Uri.parse(location),
+            // Resolves the location based on the current Uri.
+            originalUri.resolve(location),
             cookies,
           ),
         ),
       );
-    } else {
-      await cookieJar.saveFromResponse(response.realUri, cookies);
     }
   }
 }