+++ dio/CHANGELOG.md
@@ -2,7 +2,34 @@
 
 ## Unreleased
 
-*None.*
+- Make `LogInterceptor` prints in DEBUG mode (when the assertion is enabled) by default.
+- Deprecate `DioError` in favor of `DioException`.
+- Fix `IOHttpClientAdapter.onHttpClientCreate` Repeated calls
+- `IOHttpClientAdapter.onHttpClientCreate` has been deprecated and is scheduled for removal in
+  Dio 6.0.0 - Please use the replacement `IOHttpClientAdapter.createHttpClient` instead.
+
+## 5.1.2
+
+- Allow `FormData` to send a null entry value as an empty string.
+
+## 5.1.1
+
+- Revert changes to `CancelToken.cancel()` behavior, as a result the `DioError`
+  provided by the `CancelToken.cancelError` does not contain useful information
+  when the token was not used with a request.
+- Fix wrong `ListFormat` being used for comparison during encoding of `FormData`
+  and `application/x-www-form-urlencoded`, resulting in potential wrong output encoding
+  for `ListFormat.multi` and `ListFormat.multiCompatible` since Dio 4.0.x.
+- Respect `Options.listFormat` when encoding `x-www-url-encoded` content.
+
+## 5.1.0
+
+- Fix double-completion when using `connectionTimeout` on web platform.
+- Allow defining adapter methods through their constructors.
+- Fix `FormData` encoding regression for maps with dynamic keys, introduced in 5.0.3.
+- Mark several static `DioMixin` functions as `@internal`.
+- Make `DioError.stackTrace` non-nullable.
+- Ensure `DioError.stackTrace` always points to the correct call site.
 
 ## 5.0.3
 
@@ -58,7 +85,6 @@
 - [Web] support send/receive progress in web platform
 - refactor timeout logic
 - use 'arraybuffer' instead of 'blob' for xhr requests in web platform
-
 
 ## 4.0.4
 
@@ -131,7 +157,6 @@ the subsequent interceptors processing logic more finely (whether to skip them o
 2. fix #851
 3. fix #641
 
-
 ## 3.0.9 2020.2.24
 
 - Add test cases
@@ -185,9 +210,9 @@ the subsequent interceptors processing logic more finely (whether to skip them o
 
 - ~~Options.cookies~~
 
-- ~~Options.connectionTimeout~~ ；We should config connection timed out  in `BaseOptions`.  For keep-alive reasons, not every request requires a separate connection。
+- ~~Options.connectionTimeout~~ ；We should config connection timed out in `BaseOptions`. For keep-alive reasons, not every request requires a separate connection。
 
-- `Options.followRedirects`、`Options.maxRedirects`、`Response.redirects` don't make sense in Flutter Web，because redirection  can be automatically handled by browsers.
+- `Options.followRedirects`、`Options.maxRedirects`、`Response.redirects` don't make sense in Flutter Web，because redirection can be automatically handled by browsers.
 - ~~FormData.from~~，use `FormData.fromMap` instead.
 - Delete ~~Formdata.asBytes()~~、~~Formdata.asBytesAsync()~~ , use `Formdata.readAsBytes()` instead.
 - Delete ~~`UploadFileInfo`~~ class， `MultipartFile` instead.
@@ -219,7 +244,6 @@ Add `deleteOnError` parameter to `downloadUri`
 ## 2.1.9
 
 - support flutter version>=1.8 (fix #357)
-
 
 ## 2.1.8
 
@@ -254,20 +278,24 @@ First Stable version for 2.x
 ## 2.0
 
 **Refactor the Interceptors**
+
 - Support add Multiple Interceptors.
 - Add Log Interceptor
 - Add CookieManager Interceptor
 
 **API**
+
 - Support Uri
 - Support `queryParameters` for all request API
 - Modify the `get` API
 
 **Options**
+
 - Separate Options to three class: Options、BaseOptions、RequestOptions
 - Add `queryParameters` and `cookies` for BaseOptions
 
 **Adapter**
+
 - Abstract HttpClientAdapter layer.
 - Provide a DefaultHttpClientAdapter which make http requests by `dart:io:HttpClient`